add other inputs
