--- app/routes/logout.tsx
@@ -1,17 +1,0 @@
-import type {
-    ActionFunction,
-    LoaderFunction,
-} from "@remix-run/node";
-import { redirect } from "@remix-run/node";
-
-import { logout } from "~/utils/session.server";
-
-export const action: ActionFunction = async ({
-    request,
-}) => {
-    return logout(request);
-};
-
-export const loader: LoaderFunction = async () => {
-    return redirect("/");
-};
