+++ app/styles/mcq-builder.css
@@ -1,12 +1,20 @@
 @import url("https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap");
 
 .app-heading {
-  text-align: center;
   font-weight: bold;
   font-size: 30px;
+  margin: 30px 0px;
+  font-family: "Roboto";
+}
+
+.app-sub-heading {
+  text-align: center;
+  font-weight: bold;
+  font-size: 25px;
   margin: 30px 0px 0px 15px;
   font-family: "Roboto";
 }
+
 .copy-to-clipboard-container {
   align-self: flex-end;
   margin-bottom: 15px;
@@ -21,6 +29,9 @@
   padding-top: 10px;
   padding-bottom: 10px;
   z-index: 0;
+}
+.m-10 {
+  margin-bottom: 10px;
 }
 
 .ace_scroller {
@@ -47,6 +58,7 @@ table {
 
 .responsive-container {
   max-width: 1200px;
+  width: 90%;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -176,7 +188,7 @@ button {
 }
 
 .expand {
-  flex-grow: 1;
+  flex-grow: 0.2;
 }
 
 .input-field {
